A1779617_KaifengChen/faceDetection/VM/requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
autosemver==0.5.5
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
build==1.0.3
CacheControl==0.13.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dlib==19.24.2
dulwich==0.19.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.4
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.95.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.10.0
firebase-admin==6.2.0
firebase-functions==0.1.1
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
functions-framework==3.4.0
gast==0.4.0
gensim==4.3.1
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.20.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-events==0.11.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
gradio==3.15.0
grobid-client-python==0.0.5
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
html2text==2020.1.16
httpcore==0.16.3
httplib2==0.22.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
importlib-resources==5.12.0
imutils==0.5.4
inspire-utils==3.0.36
install==1.3.5
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.23.0
keras==2.13.1rc0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
libmagic==1.0
linkify-it-py==2.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.5
mistune==3.0.1
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.9
nameparser==0.5.8
nbclient==0.8.0
nbconvert==7.7.2
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==7.0.0
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
opencv-contrib-python==4.8.0.76
opt-einsum==3.3.0
orjson==3.8.7
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pdfminer==20191125
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==7.3.0
platformdirs==3.2.0
plotly==5.15.0
preshed==3.0.8
prometheus-client==0.17.1
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.4
ptyprocess==0.7.0
PuLP==2.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.6
pydub==0.25.1
Pygments==2.14.0
PyJWT==2.8.0
PyMuPDF==1.22.5
pyparsing==3.0.9
PyPDF2==1.28.6
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
rank-bm25==0.2.2
refextract==1.1.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-hub==0.13.0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.3.2
tinycss2==1.2.1
tokenizer==3.4.2
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
Unidecode==1.3.6
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.21.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
Werkzeug==3.0.0
widgetsnbextension==4.0.8
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
